+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- fix usage of query parameters with double underscore [#17](https://github.com/RedHatProductSecurity/osidb-bindings/issues/17)
 
 ## [3.3.0] - 2023-06-28
 ### Added